tools/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sushida-ocr"
version = "1.0.0"
description = "寿司打の結果画面からスコアデータを抽出するCLIツール"
authors = [
    {name = "mishio"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "click>=8.1.0",
    "numpy>=1.24.0",
    "colorama>=0.4.6",
]

[project.scripts]
sushida-ocr = "src.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

# UV用の設定
[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
